backend/requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and serialization
pydantic>=2.5.0

# LangGraph for AI workflow orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0

# AI model providers (choose one or more)
langchain-openai>=0.2.0  # For OpenAI models
langchain-ollama>=0.1.0  # For local Ollama models (optional)

# HTTP client for external API calls
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# CORS middleware
python-multipart>=0.0.6

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0